requirements.txt
tensorflow==2.15.0
keras==2.15.0
pandas==2.2.1
pylint==3.1.0
pyyaml==6.0.1
xgboost==2.0.3
ipython==8.18.1
simple-parsing==0.1.5
fastparquet==2024.2.0
scikit-learn==1.4.2
shap==0.45.0
matplotlib==3.8.4
xgboost==2.0.3
lightgbm==4.3.0
catboost==1.2.5